tools/azure-function/requirements.txt
azure-functions>=1.17.0
python-pptx>=0.6.23
msal>=1.24.0
requests>=2.31.0
